requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.8.0
asyncio-utilities>=0.1.0
python-dotenv>=1.0.0

# Repository and Git handling
GitPython>=3.1.0
gitdb>=4.0.0

# Traditional SAST tools
bandit[toml]>=1.7.0
semgrep>=1.0.0
safety>=2.0.0
pip-audit>=2.0.0

# Data processing and analysis
pydantic>=2.0.0
dataclasses-json>=0.5.0
rapidfuzz>=3.0.0

# SARIF format support
sarif-om>=1.0.0

# CLI and utilities
click>=8.0.0
rich>=13.0.0
typer>=0.9.0

# Configuration management
PyYAML>=6.0.0
configparser>=5.0.0

# Async and concurrency
asyncio>=3.4.3
aiofiles>=23.0.0

# HTTP clients and API handling
httpx>=0.24.0
tenacity>=8.0.0

# Logging and monitoring
structlog>=23.0.0
colorlog>=6.0.0

# Security and cryptography
cryptography>=41.0.0
bcrypt>=4.0.0

# Testing dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0 